Renamed cgCTL.java and cgUI.java to StudentControl and StudentMAnagment respectively. completed rules 1 through 13 of StudentControl.java. Some variable renaming still required.
