+++ src/datamanagement/Main.java
@@ -2,6 +2,6 @@ package datamanagement;
 
 public class Main {
   public static void main(String[] args) {
-    new cgCTL().execute();
+    new StudentControl().execute();
   }
 }
